pyproject.toml
[tool.poetry]
name = "portkey"
version = "0.1.0"
description = ""
authors = ["Hebert Freitas <hebertrfreitas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.35.13"
portkey-ai = "^1.7.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
boto3 = "^1.34.144"
pip-tools = "^7.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
